backend/services/user_service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.2.0